requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
pyserial==3.5
crc8==0.2.0